pyproject.toml
[project]
name = "breadbotdiscord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.6",
    "pydantic-settings>=2.10.1",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "ruff>=0.13.1",
]
discord = [
    "discord-py>=2.6.3",
]
